ch3 - clustering/requirements.txt
feedparser
attrs

# fork of PIL: https://pillow.readthedocs.io/en/stable/handbook/index.html
pillow